+++ tests/test_win32.py
@@ -14,80 +14,84 @@ ctable = os.path.join(path, "0101_20170311.ch")
 class TestWin32ExtractSACClass:
     def test_extract_sac_1(self):
         outdir = os.path.join(pwd, "test1")
-        sac = win32.extract_sac(data, ctable, outdir=outdir)
-
-        filelist = ['N.NNMH.U.SAC',
-                    'N.NNMH.N.SAC',
-                    'N.NNMH.E.SAC',
-                    'N.NGUH.U.SAC',
+        win32.extract_sac(data, ctable, outdir=outdir)
+        sac = sorted(glob.glob(os.path.join(outdir, "*.SAC")))
+        filelist = ['N.NGUH.E.SAC',
                     'N.NGUH.N.SAC',
-                    'N.NGUH.E.SAC']
+                    'N.NGUH.U.SAC',
+                    'N.NNMH.E.SAC',
+                    'N.NNMH.N.SAC',
+                    'N.NNMH.U.SAC']
         sac_to_check = [os.path.join(outdir, name) for name in filelist]
         assert sac == sac_to_check
         shutil.rmtree(outdir)
 
     def test_extract_sac_2(self):
         outdir = os.path.join(pwd, "test2")
-        sac = win32.extract_sac(data, ctable, suffix="", outdir=outdir)
-
-        filelist = ['N.NNMH.U',
-                    'N.NNMH.N',
-                    'N.NNMH.E',
-                    'N.NGUH.U',
+        win32.extract_sac(data, ctable, suffix="", outdir=outdir)
+        sac = sorted(glob.glob(os.path.join(outdir, "N.*.[ENU]")))
+        filelist = ['N.NGUH.E',
                     'N.NGUH.N',
-                    'N.NGUH.E']
+                    'N.NGUH.U',
+                    'N.NNMH.E',
+                    'N.NNMH.N',
+                    'N.NNMH.U']
         sac_to_check = [os.path.join(outdir, name) for name in filelist]
         assert sac == sac_to_check
         shutil.rmtree(outdir)
 
     def test_extract_sac_3(self):
         outdir = os.path.join(pwd, "test3")
-        sac = win32.extract_sac(data, ctable, filter_by_id='3e8?', outdir=outdir)
-
-        filelist = ['N.NNMH.U.SAC',
+        win32.extract_sac(data, ctable, filter_by_id='3e8?', outdir=outdir)
+        sac = sorted(glob.glob(os.path.join(outdir, "N.*.SAC")))
+        filelist = ['N.NNMH.E.SAC',
                     'N.NNMH.N.SAC',
-                    'N.NNMH.E.SAC']
+                    'N.NNMH.U.SAC']
         sac_to_check = [os.path.join(outdir, name) for name in filelist]
         assert sac == sac_to_check
         shutil.rmtree(outdir)
 
     def test_extract_sac_4(self):
         outdir = os.path.join(pwd, "test4")
-        sac = win32.extract_sac(data, ctable, filter_by_name='N.NG*', outdir=outdir)
-        filelist  = ['N.NGUH.U.SAC',
+        win32.extract_sac(data, ctable, filter_by_name='N.NG*', outdir=outdir)
+        sac = sorted(glob.glob(os.path.join(outdir, "N.*.SAC")))
+        filelist  = ['N.NGUH.E.SAC',
                      'N.NGUH.N.SAC',
-                     'N.NGUH.E.SAC']
+                     'N.NGUH.U.SAC']
         sac_to_check = [os.path.join(outdir, name) for name in filelist]
         assert sac == sac_to_check
         shutil.rmtree(outdir)
 
     def test_extract_sac_5(self):
         outdir = os.path.join(pwd, "test5")
-        sac = win32.extract_sac(data, ctable, filter_by_component=['N', 'E'],
-                                outdir=outdir)
-        filelist = ['N.NNMH.N.SAC',
-                    'N.NNMH.E.SAC',
+        win32.extract_sac(data, ctable, filter_by_component=['N', 'E'],
+                          outdir=outdir)
+        sac = sorted(glob.glob(os.path.join(outdir, "N.*.SAC")))
+        filelist = ['N.NGUH.E.SAC',
                     'N.NGUH.N.SAC',
-                    'N.NGUH.E.SAC']
+                    'N.NNMH.E.SAC',
+                    'N.NNMH.N.SAC']
         sac_to_check = [os.path.join(outdir, name) for name in filelist]
         assert sac == sac_to_check
         shutil.rmtree(outdir)
 
     def test_extract_sac_6(self):
         outdir = os.path.join(pwd, "test6")
-        sac, pz = win32.extract_sac(data, ctable, filter_by_component=['N', 'E'],
-                                outdir=outdir, with_pz=True)
-        filelist = ['N.NNMH.N.SAC',
-                    'N.NNMH.E.SAC',
+        win32.extract_sac(data, ctable, filter_by_component=['N', 'E'],
+                          outdir=outdir, with_pz=True)
+        sac = sorted(glob.glob(os.path.join(outdir, "N.*.SAC")))
+        filelist = ['N.NGUH.E.SAC',
                     'N.NGUH.N.SAC',
-                    'N.NGUH.E.SAC']
+                    'N.NNMH.E.SAC',
+                    'N.NNMH.N.SAC']
         sac_to_check = [os.path.join(outdir, name) for name in filelist]
         assert sac == sac_to_check
 
-        filelist = ['N.NNMH.N.SAC_PZ',
-                    'N.NNMH.E.SAC_PZ',
+        pz = sorted(glob.glob(os.path.join(outdir, "N.*.SAC_PZ")))
+        filelist = ['N.NGUH.E.SAC_PZ',
                     'N.NGUH.N.SAC_PZ',
-                    'N.NGUH.E.SAC_PZ']
+                    'N.NNMH.E.SAC_PZ',
+                    'N.NNMH.N.SAC_PZ']
         pz_to_check = [os.path.join(outdir, name) for name in filelist]
         assert pz == pz_to_check
         shutil.rmtree(outdir)