+++ Changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 0.3.4:
  - `win32.extract_sac`: skip if data not exists
+ - `win32.extract_sac`: support multiple processes to speedup, and no longer
+   return values
 
 0.3.3 (2017-03-17):
  - change ``network`` to ``NETWORK`` in ``header.py``